Add new card variation and better support columns.
