+++ blocks/cards/cards.css
@@ -1,4 +1,5 @@
 @import url('./icon.css');
+@import url('./icon-cta.css');
 @import url('./product.css');
 @import url('./profile.css');
 @import url('./cta.css');
