+++ blocks/cards/cards.js
@@ -1,6 +1,7 @@
 /* eslint-disable import/no-named-default */
-import { default as product } from './product.js';
 import { default as icon } from './icon.js';
+import { default as product } from './product.js';
+import { default as iconCta } from './icon-cta.js';
 import { default as profile } from './profile.js';
 import { default as cta } from './cta.js';
 /* eslint-enable import/no-named-default */
@@ -8,6 +9,8 @@ import { default as cta } from './cta.js';
 export default async function decorate(block) {
   if (block.classList.contains('icon')) {
     await icon(block);
+  } else if (block.classList.contains('icon-cta')) {
+    await iconCta(block);
   } else if (block.classList.contains('product')) {
     await product(block);
   } else if (block.classList.contains('profile')) {
