+++ blocks/cards/icon.js
@@ -1,44 +1,16 @@
 import { decorateIcons } from '../../scripts/lib-franklin.js';
 
-/**
- * Builds the Icons variation of the cards block.
- * @param {HTMLDivElement} block
- */
 export default async function decorate(block) {
-  [...block.children].forEach((card) => {
-    card.classList.add('card');
-    [...card.children].forEach((body) => {
-      const anchor = body.querySelector('a');
+  let count;
 
-      const content = document.createElement('a');
-      content.classList.add('card-content');
-      content.href = anchor.href;
-      content.title = anchor.title;
-
-      // Pull out the icon
-      const icon = body.querySelector('span.icon');
-      if (icon) {
-        const tmp = icon.parentElement;
-        const iconWrapper = document.createElement('div');
-        iconWrapper.classList.add('card-icon');
-        iconWrapper.append(icon);
-        tmp.remove();
-        content.append(iconWrapper);
-      } else {
-        card.classList.add('no-icon');
-      }
-
-      const link = document.createElement('div');
-      link.classList.add('card-link');
-      link.innerHTML = `<span>${anchor.textContent}</span>`;
-
-      body.classList.add('card-body');
-
-      const tmp = anchor.parentElement;
-      tmp.remove();
-
-      content.append(body);
-      card.append(content, link);
+  [...block.children].forEach((row) => {
+    if (!count) {
+      count = row.children.length;
+      block.classList.add(`cols-${count}`);
+    }
+    row.classList.add('row');
+    [...row.children].forEach((card) => {
+      card.classList.add('card');
     });
   });
   await decorateIcons(block);
