+++ styles/styles.css
@@ -302,11 +302,16 @@ main .section.cadet-gradient-background {
   );
 }
 
+main .section.center .default-content-wrapper {
+  text-align: center;
+}
+
 main .section .default-content-wrapper > p > picture {
   display: block;
   position: relative;
   height: 100%;
   width: 100%;
+  margin: 0 auto;
 }
 
 main .section .default-content-wrapper > p > picture > img {
@@ -346,7 +351,7 @@ main .section[data-layout] .layout-content-wrapper {
   gap: 3em;
 }
 
-main .section[data-layout] .layout-content-wrapper .default-content-wrapper {
+main .section[data-layout]:not(.center) .layout-content-wrapper .default-content-wrapper {
   padding: 0 24px;
   border-left: 3px solid var(--red);
 }
@@ -519,6 +524,13 @@ main .section.gray-gradient-background {
     align-items: center;
     gap: 1em;
     grid-template-columns: 1fr 3fr;
+  }
+
+  main .section[data-layout="75/25"] .layout-content-wrapper {
+    display: grid;
+    align-items: center;
+    gap: 1em;
+    grid-template-columns: 3fr 1fr;
   }
 
   main .section > div {
